changed expirySeconds to an optional int, cleaned up platform specific expiry implementation and parsing

+++ src/main/java/com/urbanairship/api/push/model/PushExpiry.java
@@ -14,10 +14,10 @@ import org.joda.time.DateTime;
 
 public class PushExpiry extends PushModelObject {
 
-    private final Optional<Long> expirySeconds;
+    private final Optional<Integer> expirySeconds;
     private final Optional<DateTime> expiryTimeStamp;
 
-    private PushExpiry(Optional<Long> expirySeconds,
+    private PushExpiry(Optional<Integer> expirySeconds,
                        Optional<DateTime> expiryTimeStamp) {
         this.expirySeconds = expirySeconds;
         this.expiryTimeStamp = expiryTimeStamp;
@@ -35,7 +35,7 @@ public class PushExpiry extends PushModelObject {
      * Get the expiry (TTL) as a long.  This is optional.
      * @return Optional<<T>ExpirySeconds</T>>
      */
-    public Optional<Long> getExpirySeconds() {
+    public Optional<Integer> getExpirySeconds() {
         return expirySeconds;
     }
 
@@ -85,7 +85,7 @@ public class PushExpiry extends PushModelObject {
     }
 
     public static class Builder {
-        private Long expirySeconds = null;
+        private Integer expirySeconds = null;
         private DateTime expiryTimeStamp = null;
 
         private Builder() { }
@@ -95,7 +95,7 @@ public class PushExpiry extends PushModelObject {
          * @param value Long
          * @return Long
          **/
-        public Builder setExpirySeconds(long value) {
+        public Builder setExpirySeconds(int value) {
             this.expirySeconds = value;
             return this;
         }
